Repeated measure sample data

A repeated measure sample data in which 3 individuals have been analysed. In the "id" column, a unique number is assigned to each individual, e.g. all the three number 1 indicate one individual. samples are taken or measurements are scored over different time points (time column) from each individual.

Usage

data_repeated_measure_1

Format

A data frame with 9 observations and 6 variables:

id

experimental factor

time

time course levels

E_target

Amplification efficiency of target gene

Ct_target

Ct of target gene

E_ref

Amplification efficiency of reference gene

Ct_ref

Ct of reference gene

Source

NA