Relative expression (\(\Delta C_T\) method) analysis using ANOVA

Analysis of variance of relative expression (\(\Delta C_T\) method) values for all factor level combinations in the experiment in which the expression level of a reference gene is used as normalizer.

qpcrANOVARE(x, numberOfrefGenes, block, alpha = 0.05, adjust = "none")

Arguments

x

a data frame consisting of condition columns, target gene efficiency (E), target Gene Ct, reference gene efficiency and reference gene Ct values, respectively. Each Ct in the data frame is the mean of technical replicates. Complete amplification efficiencies of 2 was assumed in the example data for all wells but the calculated efficienies can be used instead. NOTE: Each line belongs to a separate individual reflecting a non-repeated measure experiment). See vignette, section "data structure and column arrangement" for details.

numberOfrefGenes

number of reference genes (1 or 2). Up to two reference genes can be handled.

block

column name of the blocking factor (for correct column arrangement see example data.). When a qPCR experiment is done in multiple qPCR plates, variation resulting from the plates may interfere with the actual amount of gene expression. One solution is to conduct each plate as a complete randomized block so that at least one replicate of each treatment and control is present on a plate. Block effect is usually considered as random and its interaction with any main effect is not considered.

alpha

significance level

adjust

method for adjusting p-values

Value

A list with 4 elements:

Final_data

The row data plus weighed delta Ct (wDCt) values.

lm

The output of linear model analysis including ANOVA tables

ANOVA

ANOVA table based on CRD

Result

The result table including treatments and factors, RE (Relative Expression), LCL, UCL, letter display for pair-wise comparisons and standard error with the lower and upper limits.

Details

The qpcrANOVARE function performs analysis of variance (ANOVA) of relative expression (RE) values for all factor level combinations as treatments using the expression level of a reference gene is used as normalizer. To get a reliable result, the expression of the reference gene needs to be constant across all test samples and it expression should not be affected by the experimental treatment under study.

References

Livak, Kenneth J, and Thomas D Schmittgen. 2001. Analysis of Relative Gene Expression Data Using Real-Time Quantitative PCR and the Double Delta CT Method. Methods 25 (4). doi:10.1006/meth.2001.1262.

Ganger, MT, Dietz GD, and Ewing SJ. 2017. A common base method for analysis of qPCR data and the application of simple blocking in qPCR experiments. BMC bioinformatics 18, 1-11.

Yuan, Joshua S, Ann Reed, Feng Chen, and Neal Stewart. 2006. Statistical Analysis of Real-Time PCR Data. BMC Bioinformatics 7 (85). doi:10.1186/1471-2105-7-85.

Author

Ghader Mirzaghaderi

Examples

# If the data include technical replicates, means of technical replicates
# should be calculated first using meanTech function.
# Applying ANOVA
qpcrANOVARE(data_3factor, numberOfrefGenes = 1, block = NULL)
#> Analysis of Variance Table
#> 
#> Response: wDCt
#>           Df Sum Sq Mean Sq F value    Pr(>F)    
#> T         11 94.001  8.5456  29.188 3.248e-11 ***
#> Residuals 24  7.027  0.2928                      
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Relative expression table
#>    Type Conc SA     RE    LCL    UCL     se Lower.se Upper.se  letters
#> 1     S    H A2 5.1934 3.3217 8.1197 0.1309   4.7429   5.6867  a      
#> 2     S    H A1 2.9690 1.8990 4.6420 0.0551   2.8577   3.0846  ab     
#> 3     R    H A2 1.7371 1.1110 2.7159 0.0837   1.6392   1.8409   bc    
#> 4     S    L A2 1.5333 0.9807 2.3973 0.0865   1.4441   1.6280    c    
#> 5     R    H A1 0.9885 0.6323 1.5455 0.0841   0.9325   1.0478    cd   
#> 6     S    L A1 0.7955 0.5088 1.2438 0.2128   0.6864   0.9219     d   
#> 7     S    M A2 0.7955 0.5088 1.2438 0.2571   0.6656   0.9507     d   
#> 8     R    M A1 0.6271 0.4011 0.9804 0.4388   0.4626   0.8500     de  
#> 9     S    M A1 0.4147 0.2652 0.6483 0.2540   0.3478   0.4945      ef 
#> 10    R    M A2 0.3150 0.2015 0.4925 0.2890   0.2578   0.3849       f 
#> 11    R    L A1 0.2852 0.1824 0.4459 0.0208   0.2811   0.2893       f 
#> 12    R    L A2 0.0641 0.0410 0.1002 0.8228   0.0362   0.1134        g

qpcrANOVARE(data_2factorBlock, block = "Block", numberOfrefGenes = 1)
#> Analysis of Variance Table
#> 
#> Response: wDCt
#>           Df  Sum Sq Mean Sq F value    Pr(>F)    
#> block      1  0.0072  0.0072  0.0425    0.8404    
#> T          5 20.5489  4.1098 24.1712 1.377e-05 ***
#> Residuals 11  1.8703  0.1700                      
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Relative expression table
#>   factor1 factor2     RE    LCL    UCL     se Lower.se Upper.se letters
#> 1       S     0.5 2.9545 2.0470 4.2644 0.0551   2.8438   3.0695    a   
#> 2       R     0.5 0.9837 0.6815 1.4198 0.0841   0.9280   1.0427     b  
#> 3       S       0 0.7916 0.5485 1.1426 0.2128   0.6830   0.9174     b  
#> 4       R    0.25 0.6240 0.4323 0.9006 0.4388   0.4604   0.8458     bc 
#> 5       S    0.25 0.4126 0.2859 0.5956 0.2540   0.3460   0.4920      cd
#> 6       R       0 0.2838 0.1966 0.4096 0.0208   0.2797   0.2879       d