Fold change (\(\Delta \Delta C_T\) method) analysis of target genes using t-test

t.test based analysis of the fold change expression for any number of target genes.

qpcrTTEST(x, numberOfrefGenes, paired = FALSE, var.equal = TRUE)

Arguments

x

a data frame of 4 columns including Conditions, E (efficiency), Gene and Ct values (see examples below). Biological replicates needs to be equal for all Genes. Each Ct value is the mean of technical replicates. Complete amplification efficiencies of 2 is assumed here for all wells but the calculated efficienies can be used instead. See vignette for details about "data structure and column arrangement".

numberOfrefGenes

number of reference genes. Up to two reference genes can be handled.

paired

a logical indicating whether you want a paired t-test.

var.equal

a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

Value

A list of two elements:

Row_data

The row data including Genes and weighed delta Ct (wDCt) values.

Result

Output table including the Fold Change values, lower and upper confidence interval, pvalue and standard error with the lower and upper limits.

For more information about the test procedure and its arguments, refer t.test, and lm. If the residuals of the model do not follow normal distribution and variances between the two groups are not homoGene, wilcox.test procedure may be concidered

Details

The qpcrTTEST function applies a t.test based analysis to calculate fold change (\(\Delta \Delta C_T\) method) expression and returns related statistics for any number of target genes that have been evaluated under control and treatment conditions. Sampling may be paired or unpaired. One or two reference genes can be used. Unpaired and paired samples are commonly analyzed using unpaired and paired t-test, respectively. NOTE: Paired samples in quantitative PCR refer to two sample data that are collected from one set of individuals at two different conditions, for example before and after a treatment or at two different time points. While for unpaired samples, two sets of individuals are used: one under untreated and the other set under treated condition. Paired samples allow to compare gene expression changes within the same individual, reducing inter-individual variability.

References

Livak, Kenneth J, and Thomas D Schmittgen. 2001. Analysis of Relative Gene Expression Data Using Real-Time Quantitative PCR and the Double Delta CT Method. Methods 25 (4). doi:10.1006/meth.2001.1262.

Ganger, MT, Dietz GD, and Ewing SJ. 2017. A common base method for analysis of qPCR data and the application of simple blocking in qPCR experiments. BMC bioinformatics 18, 1-11.

Yuan, Joshua S, Ann Reed, Feng Chen, and Neal Stewart. 2006. Statistical Analysis of Real-Time PCR Data. BMC Bioinformatics 7 (85). doi:10.1186/1471-2105-7-85.

Author

Ghader Mirzaghaderi

Examples

# See the sample data structure
data_ttest
#>    Condition    Gene E    Ct
#> 1    control C2H2-26 2 31.26
#> 2    control C2H2-26 2 31.01
#> 3    control C2H2-26 2 30.97
#> 4  treatment C2H2-26 2 32.65
#> 5  treatment C2H2-26 2 32.03
#> 6  treatment C2H2-26 2 32.40
#> 7    control C2H2-01 2 31.06
#> 8    control C2H2-01 2 30.41
#> 9    control C2H2-01 2 30.97
#> 10 treatment C2H2-01 2 28.85
#> 11 treatment C2H2-01 2 28.93
#> 12 treatment C2H2-01 2 28.90
#> 13   control C2H2-12 2 28.50
#> 14   control C2H2-12 2 28.40
#> 15   control C2H2-12 2 28.80
#> 16 treatment C2H2-12 2 27.90
#> 17 treatment C2H2-12 2 28.00
#> 18 treatment C2H2-12 2 27.90
#> 19   control     ref 2 28.87
#> 20   control     ref 2 28.42
#> 21   control     ref 2 28.53
#> 22 treatment     ref 2 28.31
#> 23 treatment     ref 2 29.14
#> 24 treatment     ref 2 28.63

# Getting t.test results
qpcrTTEST(data_ttest,
   paired = FALSE,
   var.equal = TRUE,
   numberOfrefGenes = 1)
#> $Raw_data
#>    Var2  wDCt
#> 1     1  2.39
#> 2     1  2.59
#> 3     1  2.44
#> 4     1  4.34
#> 5     1  2.89
#> 6     1  3.77
#> 7     2  2.19
#> 8     2  1.99
#> 9     2  2.44
#> 10    2  0.54
#> 11    2 -0.21
#> 12    2  0.27
#> 13    3 -0.37
#> 14    3 -0.02
#> 15    3  0.27
#> 16    3 -0.41
#> 17    3 -1.14
#> 18    3 -0.73
#> 
#> $Result
#>      Gene     FC    LCL    UCL pvalue     se Lower.se Upper.se
#> 1 C2H2-26 0.4373 0.1926 0.9927 0.0488 0.4218   0.3264   0.5858
#> 2 C2H2-01 4.0185 2.4598 6.5649 0.0014 0.2193   3.4518   4.6782
#> 3 C2H2-12 1.6472 0.9595 2.8279 0.0624 0.2113   1.4228   1.9070
#> 

qpcrTTEST(Taylor_etal2019, 
          numberOfrefGenes = 2, 
          var.equal = TRUE)
#> $Raw_data
#>   Var2  wDCt
#> 1    1 2.390
#> 2    1 2.585
#> 3    1 2.405
#> 4    1 4.330
#> 5    1 2.890
#> 6    1 3.770
#> 
#> $Result
#>   Gene     FC    LCL    UCL pvalue     se Lower.se Upper.se
#> 1 DER5 0.4343 0.1921 0.9816 0.0469 0.4191   0.3248   0.5807
#>